fix can rx api
